Move Room and Level from core to models
