+++ roguelike-game/src/main/java/core/GameState.java
@@ -1,5 +1,7 @@
 package core;
 
+import models.Level;
+import models.Room;
 import models.Player;
 import repositories.LevelRepository;
 import lib.MapGenerator;
